Migrate to standalone Parlot source generation
Use Parlot.SourceGenerator 2.0.0-preview-763 without a production Parlot dependency. Preserve per-call options, culture, cancellation, and legacy framework support; refresh Fluent/generated benchmarks and documentation.

BREAKING CHANGE: Remove GetOrCreateExpressionParser and Parlot inheritance from LogicalExpressionParseContext; use the direct Parse APIs instead.
