+++ nuget.config
@@ -10,6 +10,7 @@
     </packageSource>
     <packageSource key="parlot-preview">
       <package pattern="Parlot" />
+      <package pattern="Parlot.SourceGenerator" />
     </packageSource>
   </packageSourceMapping>
 </configuration>
