+++ src/NCalc.Core/Factories/LogicalExpressionFactory.cs
@@ -36,7 +36,6 @@ public sealed class LogicalExpressionFactory(ILogger<LogicalExpressionFactory>? 
 
     public static LogicalExpression Create(string expression, LogicalExpressionParserOptions? options = null, CultureInfo? cultureInfo = null, CancellationToken cancellationToken = default)
     {
-        var parserContext = new LogicalExpressionParseContext(expression, options ?? new LogicalExpressionParserOptions(), cancellationToken);
-        return LogicalExpressionParser.Parse(parserContext, cultureInfo);
+        return LogicalExpressionParser.Parse(expression, options, cultureInfo, cancellationToken);
     }
 }
