--- src/NCalc.Parser/LogicalExpressionParserCacheKey.cs
@@ -1,19 +1,0 @@
-namespace NCalc;
-
-internal readonly record struct LogicalExpressionParserCacheKey(
-    bool AllowCharValues,
-    bool DisallowSingleEquals,
-    FloatingPointNumberType FloatingPointNumberType,
-    IntegerNumberType IntegerNumberType,
-    ArgumentSeparator ArgumentSeparator)
-{
-    public LogicalExpressionParserCacheKey(LogicalExpressionParserOptions options)
-        : this(
-            options.AllowCharValues,
-            options.DisallowSingleEquals,
-            options.FloatingPointNumberType,
-            options.IntegerNumberType,
-            options.ArgumentSeparator)
-    {
-    }
-}
