+++ test/NCalc.Benchmarks/FluentExpressionParser.cs
@@ -1,0 +1,11 @@
+using Parlot.Fluent;
+
+namespace NCalc.Benchmarks;
+
+internal static partial class FluentExpressionParser
+{
+    internal static Parser<LogicalExpression> Create(LogicalExpressionParserOptions options, CultureInfo culture)
+    {
+        return CreateParserGrammar(options, culture);
+    }
+}
