+++ test/NCalc.Benchmarks/ParserGenerationBenchmark.cs
@@ -1,9 +1,9 @@
-using System.Reflection;
 using BenchmarkDotNet.Attributes;
 using BenchmarkDotNet.Configs;
 using BenchmarkDotNet.Jobs;
 using BenchmarkDotNet.Order;
 using NCalc.Extensions;
+using Parlot;
 using Parlot.Fluent;
 
 namespace NCalc.Benchmarks;
@@ -24,80 +24,55 @@ public class ParserGenerationBenchmark
     private LogicalExpressionParserOptions ParserOptions { get; } = new();
     private CultureInfo ParserCulture { get; } = CultureInfo.InvariantCulture;
 
-    private Func<LogicalExpressionParserOptions, CultureInfo, Parser<LogicalExpression>> DynamicFactory { get; set; }
-    private Func<LogicalExpressionParserOptions, CultureInfo, Parser<LogicalExpression>> GeneratedFactory { get; set; }
-    private Parser<LogicalExpression> DynamicParser { get; set; }
-    private Parser<LogicalExpression> GeneratedParser { get; set; }
+    private Parser<LogicalExpression> FluentParser { get; set; }
 
     [GlobalSetup]
     public void Setup()
     {
-        // The grammar delegate bypasses interception; the wrapper delegate constructs a fresh generated parser.
-        DynamicFactory = CreateFactory("CreateParserGrammar");
-        GeneratedFactory = CreateFactory("CreateExpressionParser");
-        DynamicParser = ConstructDynamic();
-        GeneratedParser = ConstructGenerated();
+        FluentParser = FluentExpressionParser.Create(ParserOptions, ParserCulture);
 
-        if (DynamicParser.GetType().Assembly != typeof(Parser<LogicalExpression>).Assembly ||
-            GeneratedParser.GetType().DeclaringType != typeof(LogicalExpressionParser))
+        if (FluentParser.GetType().Assembly != typeof(Parser<LogicalExpression>).Assembly ||
+            typeof(LogicalExpressionParser).Assembly.GetReferencedAssemblies().Any(reference => reference.Name == "Parlot"))
         {
-            throw new InvalidOperationException("The comparison requires a runtime combinator graph and a source-generated parser.");
+            throw new InvalidOperationException("The comparison requires a Fluent graph and a dependency-free generated parser.");
         }
 
         foreach (var text in new[] { SimpleExpression, AdvancedExpression })
         {
-            if (Parse(DynamicParser, text).ToExpressionString() != Parse(GeneratedParser, text).ToExpressionString())
+            if (ParseFluent(text).ToExpressionString() != ParseGenerated(text).ToExpressionString())
             {
-                throw new InvalidOperationException("The dynamic and generated parsers produced different expressions.");
+                throw new InvalidOperationException("The Fluent and generated parsers produced different expressions.");
             }
         }
     }
 
     [Benchmark(Baseline = true)]
-    [BenchmarkCategory("Construction")]
-    public Parser<LogicalExpression> ConstructDynamic() => DynamicFactory(ParserOptions, ParserCulture);
-
-    [Benchmark]
-    [BenchmarkCategory("Construction")]
-    public Parser<LogicalExpression> ConstructGenerated() => GeneratedFactory(ParserOptions, ParserCulture);
-
-    [Benchmark(Baseline = true)]
     [BenchmarkCategory("Simple")]
-    public LogicalExpression ParseSimpleDynamic() => Parse(DynamicParser, SimpleExpression);
+    public LogicalExpression ParseSimpleFluent() => ParseFluent(SimpleExpression);
 
     [Benchmark]
     [BenchmarkCategory("Simple")]
-    public LogicalExpression ParseSimpleGenerated() => Parse(GeneratedParser, SimpleExpression);
+    public LogicalExpression ParseSimpleGenerated() => ParseGenerated(SimpleExpression);
 
     [Benchmark(Baseline = true)]
     [BenchmarkCategory("Advanced")]
-    public LogicalExpression ParseAdvancedDynamic() => Parse(DynamicParser, AdvancedExpression);
+    public LogicalExpression ParseAdvancedFluent() => ParseFluent(AdvancedExpression);
 
     [Benchmark]
     [BenchmarkCategory("Advanced")]
-    public LogicalExpression ParseAdvancedGenerated() => Parse(GeneratedParser, AdvancedExpression);
+    public LogicalExpression ParseAdvancedGenerated() => ParseGenerated(AdvancedExpression);
 
-    private static Func<LogicalExpressionParserOptions, CultureInfo, Parser<LogicalExpression>> CreateFactory(string name)
+    private LogicalExpression ParseFluent(string text)
     {
-        var factory = typeof(LogicalExpressionParser).GetMethod(
-            name,
-            BindingFlags.Static | BindingFlags.NonPublic,
-            binder: null,
-            types: [typeof(LogicalExpressionParserOptions), typeof(CultureInfo)],
-            modifiers: null)
-            ?? throw new InvalidOperationException($"The parser factory '{name}' was not found.");
-
-        return factory.CreateDelegate<Func<LogicalExpressionParserOptions, CultureInfo, Parser<LogicalExpression>>>();
-    }
-
-    private LogicalExpression Parse(Parser<LogicalExpression> parser, string text)
-    {
-        var context = new LogicalExpressionParseContext(text, ParserOptions, CancellationToken.None);
-        if (parser.TryParse(context, out var result, out var error))
+        var context = new ParseContext(new Scanner(text), useNewLines: false, disableLoopDetection: true, CancellationToken.None);
+        if (FluentParser.TryParse(context, out var result, out var error))
         {
             return result;
         }
 
         throw new InvalidOperationException($"Failed to parse the benchmark expression: {error?.Message}");
     }
+
+    private LogicalExpression ParseGenerated(string text) =>
+        LogicalExpressionParser.Parse(text, ParserOptions, ParserCulture, CancellationToken.None);
 }
